+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## master
+* feat: FaultInjection entity, support in API and `kumactl` 
+  [#643](https://github.com/Kong/kuma/pull/643)
 * chore: merge mTLS and CA status into one column
   [#637](https://github.com/Kong/kuma/pull/637)
 * fix: `kumactl apply -v ...` support dots in variables name